+++ .typescript-declarations/cinnamon/ui/boxpointer.d.ts
@@ -1,5 +1,0 @@
-declare namespace imports.ui.boxpointer {
-    class BoxPointer {
-
-    }
-}