+++ sound150@claudiux/files/sound150@claudiux/6.6/applet.js
@@ -36,6 +36,7 @@ const Gettext = imports.gettext;
 const Extension = imports.ui.extension;
 const Pango = imports.gi.Pango;
 const ModalDialog = imports.ui.modalDialog;
+const Signals = imports.signals;
 
 
 let HtmlEncodeDecode = require("./lib/htmlEncodeDecode");
@@ -55,6 +56,7 @@ const {
 const { del_song_arts } = require("./lib/del_song_arts");
 
 const { VolumeSlider } = require("./lib/volumeSlider");
+const { BalanceSlider } = require("./lib/balanceSlider");
 const { ControlButton } = require("./lib/controlButton");
 const { StreamMenuSection, Player, MediaPlayerLauncher, Seeker } = require("./lib/s150PopupMenu");
 
@@ -220,11 +222,9 @@ class Sound150Applet extends Applet.TextIconApplet {
         this.alreadyCalledBysetAppletTooltip = false;
 
         this.real_ui_scale = 1.0;
-        this.menuWidth = this.userMenuWidth; //520;
         Util.spawnCommandLineAsyncIO(PATH2SCRIPTS + "/get-real-scale.py", (stdout, stderr, exitCode) => {
             if (exitCode === 0) {
                 this.real_ui_scale = parseFloat(stdout);
-                this.menuWidth = Math.round(this.userMenuWidth * this.real_ui_scale);
             }
         }, {});
 
@@ -273,10 +273,10 @@ class Sound150Applet extends Applet.TextIconApplet {
         this._chooseActivePlayerItem = new PopupMenu.PopupSubMenuMenuItem(_("Choose player controls"));
 
         this.settings = new Settings.AppletSettings(this, UUID, this.instanceId);
-        this.settings.bind("userMenuWidth", "userMenuWidth", (value) => {
-            this.menuWidth = Math.round(value * this.real_ui_scale);
-        });
-        this.menuWidth = Math.round(this.userMenuWidth * this.real_ui_scale);
+        
+        this.settings.bind("userMenuWidth", "popup_width");
+        this.settings.bind("userMenuHeight", "popup_height");
+        
         this.settings.bind("ignoredInputDevices", "ignoredInputDevices", () => {this._on_reload_this_applet_pressed();});
         this.settings.bind("ignoredOutputDevices", "ignoredOutputDevices", () => {this._on_reload_this_applet_pressed();});
         this.settings.bind("runAsync", "runAsync");
@@ -379,6 +379,15 @@ class Sound150Applet extends Applet.TextIconApplet {
         this.settings.bind("keySwitchPlayer", "keySwitchPlayer", () => {
             this._setKeybinding()
         });
+        this.settings.bind("keyBalanceLeft", "keyBalanceLeft", () => {
+            this._setKeybinding()
+        });
+        this.settings.bind("keyBalanceRight", "keyBalanceRight", () => {
+            this._setKeybinding()
+        });
+        this.settings.bind("keyBalanceCenter", "keyBalanceCenter", () => {
+            this._setKeybinding()
+        });
 
         this.settings.bind("magneticOn", "magneticOn", () => this._on_sound_settings_change());
         this.settings.bind("magnetic25On", "magnetic25On", () => this._on_sound_settings_change());
@@ -414,6 +423,7 @@ class Sound150Applet extends Applet.TextIconApplet {
         this.settings.bind("volumeSoundEnabled", "volumeSoundEnabled", this.on_volumeSoundEnabled_changed);
         this.settings.bind("maxVolume", "maxVolume", (value) => this._on_maxVolume_changed(value));
 
+        this.settings.bind("balance", "balance");
         this.settings.bind("volume", "volume");
 
         this.old_volume = this.volume;
@@ -507,6 +517,13 @@ class Sound150Applet extends Applet.TextIconApplet {
         this.menuManager = new PopupMenu.PopupMenuManager(this);
         this.menu = new Applet.AppletPopupMenu(this, this.orientation);
         this.menuManager.addMenu(this.menu);
+        this._resizer = new Applet.PopupResizeHandler(
+            this.menu.actor,
+            () => this.orientation,
+            (w,h) => this._onBoxResized(w,h),
+            () => this.popup_width * this.real_ui_scale,
+            () => this.popup_height * this.real_ui_scale
+        );
 
         this.set_applet_icon_symbolic_name("audio-x-generic");
 
@@ -598,6 +615,8 @@ class Sound150Applet extends Applet.TextIconApplet {
         this._output = null;
         this._outputMutedId = null;
         this._outputIcon = "audio-volume-muted-symbolic";
+        
+        this._channelMap = null;
 
         this._input = null;
         this._inputMutedId = null;
@@ -712,6 +731,17 @@ class Sound150Applet extends Applet.TextIconApplet {
 
         this._sound_settings.connect("changed::" + OVERAMPLIFICATION_KEY, () => this._on_sound_settings_change());
     }
+    
+    _onBoxResized(width, height) {
+        this.menu.actor.set_width(width);
+        this.menu.actor.set_height(height);
+        if (!this._resizer.resizingInProgress) {
+           //~ this.settings.popup_width = width / this.real_ui_scale;
+           //~ this.settings.popup_height = height / this.real_ui_scale;
+           this.popup_width = Math.max( Math.round(width / this.real_ui_scale), 450 );
+           this.popup_height = Math.round(height / this.real_ui_scale);
+        }
+    }
 
     monitor_icon_dir() {
         this.unmonitor_icon_dir();
@@ -743,7 +773,8 @@ class Sound150Applet extends Applet.TextIconApplet {
      unmonitor_art_dir() {
         if (this.artsMonitor == null || this.artsMonitorId == null || this.artsMonitor.is_cancelled()) return;
         
-        this.artsMonitor.disconnect(this.iconsMonitorId);
+        //~ this.artsMonitor.disconnect(this.iconsMonitorId);
+        this.artsMonitor.disconnect(this.artsMonitorId);
         if (! this.artsMonitor.is_cancelled() )
             this.artsMonitor.cancel();
         this.artsMonitor = null;
@@ -1084,8 +1115,43 @@ class Sound150Applet extends Applet.TextIconApplet {
         }
         return commandline;
     }
+    
+    _balanceLeft() {
+        let bal = this.balance;
+        this.balance = Math.max(
+            0,
+            bal - 0.05
+        );
+        if (this._channelMap)
+            this._channelMap.set_balance(2 * this.balance - 1);
+    }
+    
+    _balanceRight() {
+        let bal = this.balance;
+        this.balance = Math.min(
+            1,
+            bal + 0.05
+        );
+        if (this._channelMap)
+            this._channelMap.set_balance(2 * this.balance - 1);
+    }
+    
+    _balanceCenter() {
+        this.balance = 0.5;
+        if (this._channelMap)
+            this._channelMap.set_balance(0);
+    }
 
     _setKeybinding() {
+        Main.keybindingManager.addHotKey("balance-left-" + this.instance_id, this.keyBalanceLeft, () => {
+            this._balanceLeft()
+        });
+        Main.keybindingManager.addHotKey("balance-right-" + this.instance_id, this.keyBalanceRight, () => {
+            this._balanceRight()
+        });
+        Main.keybindingManager.addHotKey("balance-center-" + this.instance_id, this.keyBalanceCenter, () => {
+            this._balanceCenter()
+        });
         Main.keybindingManager.addHotKey("sound-open-" + this.instance_id, this.keyOpen, () => {
             this._openMenu()
         });
@@ -1267,8 +1333,7 @@ class Sound150Applet extends Applet.TextIconApplet {
 
     on_applet_added_to_panel() {
         this.themeNode = null;
-        //~ this.set_applet_label(this._applet_label.get_text());
-        this.menu.actor.set_width(this.menuWidth);
+        this.menu.actor.set_width(this.popup_width);
         this.title_text_old = "";
         this.startingUp = true;
         if (this._playerctl)
@@ -1359,6 +1424,9 @@ class Sound150Applet extends Applet.TextIconApplet {
             this.volume = this.old_volume;
         }
 
+        Main.keybindingManager.removeHotKey("balance-left-" + this.instance_id);
+        Main.keybindingManager.removeHotKey("balance-right-" + this.instance_id);
+        Main.keybindingManager.removeHotKey("balance-center-" + this.instance_id);
         Main.keybindingManager.removeHotKey("sound-open-" + this.instance_id);
         Main.keybindingManager.removeHotKey("switch-player-" + this.instance_id);
         try {
@@ -1440,11 +1508,12 @@ class Sound150Applet extends Applet.TextIconApplet {
             this._remove_OsdWithNumberATJosephMcc_button.actor.show();
         else
             this._remove_OsdWithNumberATJosephMcc_button.actor.hide();
-        this.menu.actor.set_width(this.menuWidth);
+        this._balanceSection._onValueInit();
+        this.menu.actor.set_width(this.popup_width);
     }
 
     _openMenu() {
-        this.menu.actor.set_width(this.menuWidth);
+        this.menu.actor.set_width(this.popup_width);
         this.menu.toggle(true);
     }
 
@@ -1670,7 +1739,7 @@ class Sound150Applet extends Applet.TextIconApplet {
             icon_name += "-symbolic";
             this._outputIcon = icon_name;
             let icon = Gio.Icon.new_for_string(icon_name);
-            //~ if (! this.keepAlbumArtIcon) 
+            if (! this.keepAlbumArtIcon) // Is it enough????
                 this.set_applet_icon_symbolic_name(icon_name);
             let _bar_level = null;
             let _volume_str = "";
@@ -1840,8 +1909,8 @@ class Sound150Applet extends Applet.TextIconApplet {
                 } else {
                     //DON'T change the icon:
                     if (this._playerIcon && this._playerIcon[0] != this.oldPlayerIcon0 || !this._iconTimeoutId) {
-                        //~ if (! this.keepAlbumArtIcon) this.set_applet_icon_symbolic_name(this._playerIcon[0]);
-                        this.set_applet_icon_symbolic_name(this._playerIcon[0]);
+                        if (! this.keepAlbumArtIcon) this.set_applet_icon_symbolic_name(this._playerIcon[0]);
+                        //~ this.set_applet_icon_symbolic_name(this._playerIcon[0]);
                         this.oldPlayerIcon0 = this._playerIcon[0];
                     }
                 }
@@ -2308,6 +2377,11 @@ class Sound150Applet extends Applet.TextIconApplet {
 
         this.menu.addMenuItem(this._outputVolumeSection);
         this.menu.addMenuItem(this._inputVolumeSection);
+        
+        this.menu.addMenuItem(new PopupMenu.PopupSeparatorMenuItem());
+        
+        this._balanceSection = new BalanceSlider(this);
+        this.menu.addMenuItem(this._balanceSection);
 
         this.menu.addMenuItem(new PopupMenu.PopupSeparatorMenuItem());
 
@@ -2561,6 +2635,7 @@ class Sound150Applet extends Applet.TextIconApplet {
         }
         this._output = this._control.get_default_sink();
         if (this._output) {
+            this._channelMap = this._output.get_channel_map();
             this._outputVolumeSection.connectWithStream(this._output);
             this._outputMutedId = this._output.connect("notify::is-muted", (...args) => this._mutedChanged(...args, "_output"));
             this._mutedChanged(null, null, "_output");
@@ -2748,6 +2823,7 @@ class Sound150Applet extends Applet.TextIconApplet {
     _removeAllStreams() {
         for (let i = 0, l = this._streams.length; i < l; ++i) {
             let stream = this._streams[i];
+            if (! stream) continue;
             if (stream.item) {
                 try { stream.item.destroy() } catch(e) {}
             }
@@ -3015,6 +3091,7 @@ class Sound150Applet extends Applet.TextIconApplet {
         return GLib.file_test(ALBUMART_ON, GLib.FileTest.EXISTS);
     }
 }
+Signals.addSignalMethods(Sound150Applet.prototype);
 
 function main(metadata, orientation, panel_height, instanceId) {
     return new Sound150Applet(metadata, orientation, panel_height, instanceId);