+++ sound150@claudiux/files/sound150@claudiux/CHANGELOG.md
@@ -1,3 +1,10 @@
+### v14.1.0~20260212
+  * Resize the menu by grabbing its edges with the mouse.
+
+### v14.0.0~20260211
+  * Adds a Balance slider in the menu + keybindings.
+  * Fixes [#8052](https://github.com/linuxmint/cinnamon-spices-applets/issues/8052)
+
 ### v13.10.0~20260121
   * Try to avoid advertisements by detecting certain words in the title, defined by the user.